requirements.txt
pandas==2.3.0
tqdm==4.66.4
tushare==1.4.21
scipy==1.14.1
pyyaml>=6.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
